+++ praw/exceptions.py
@@ -5,6 +5,7 @@ wrong on the server side, and :class:`.ClientException` when something goes
 wrong on the client side. Both of these classes extend :class:`.PRAWException`.
 
 """
+from typing import Optional
 
 
 class PRAWException(Exception):
@@ -14,7 +15,7 @@ class PRAWException(Exception):
 class APIException(PRAWException):
     """Indicate exception that involve responses from Reddit's API."""
 
-    def __init__(self, error_type, message, field):
+    def __init__(self, error_type: str, message: str, field: Optional[str]):
         """Initialize an instance of APIException.
 
         :param error_type: The error type set on Reddit's end.