+++ praw/models/base.py
@@ -1,5 +1,8 @@
 """Provide the PRAWBase superclass."""
 from copy import deepcopy
+from typing import Any, Dict, Optional, TypeVar
+
+Reddit = TypeVar("Reddit")
 
 
 class PRAWBase:
@@ -20,7 +23,7 @@ class PRAWBase:
         argument_dict[key] = value
 
     @classmethod
-    def parse(cls, data, reddit):
+    def parse(cls, data: Dict[str, Any], reddit: Reddit) -> Any:
         """Return an instance of ``cls`` from ``data``.
 
         :param data: The structured data.
@@ -29,7 +32,7 @@ class PRAWBase:
         """
         return cls(reddit, _data=data)
 
-    def __init__(self, reddit, _data):
+    def __init__(self, reddit: Reddit, _data: Optional[Dict[str, Any]]):
         """Initialize a PRAWModel instance.
 
         :param reddit: An instance of :class:`.Reddit`.