+++ tests/conftest.py
@@ -1,13 +1,15 @@
 """Prepare py.test."""
+from base64 import b64encode
+from functools import wraps
 import json
 import os
 import socket
 import sys
-import time
-from base64 import b64encode
 from sys import platform
+import time
 
 import betamax
+from betamax.cassette.cassette import dispatch_hooks, Cassette
 import pytest
 from betamax_serializers import pretty_json
 
@@ -67,9 +69,11 @@ placeholders = {
     x: env_default(x)
     for x in (
         "auth_code client_id client_secret password redirect_uri "
-        "test_subreddit user_agent username"
+        "test_subreddit user_agent username refresh_token"
     ).split()
 }
+
+
 placeholders["basic_auth"] = b64_string(
     "{}:{}".format(placeholders["client_id"], placeholders["client_secret"])
 )
@@ -84,6 +88,28 @@ with betamax.Betamax.configure() as config:
         if key == "password":
             value = quote_plus(value)
         config.define_cassette_placeholder("<{}>".format(key.upper()), value)
+
+
+def add_init_hook(original_init):
+    """Wrap an __init__ method to also call some hooks."""
+
+    @wraps(original_init)
+    def wrapper(self, *args, **kwargs):
+        original_init(self, *args, **kwargs)
+        dispatch_hooks("after_init", self)
+
+    return wrapper
+
+
+Cassette.__init__ = add_init_hook(Cassette.__init__)
+
+
+def init_hook(cassette):
+    if cassette.is_recording():
+        pytest.set_up_record()  # dynamically defined in __init__.py
+
+
+Cassette.hooks["after_init"].append(init_hook)
 
 
 class Placeholders: