+++ tests/integration/__init__.py
@@ -1,7 +1,9 @@
 """PRAW Integration test suite."""
-import pytest
+
 from betamax import Betamax
 from praw import Reddit
+import pytest
+import requests
 
 from .. import PrawTest
 
@@ -14,6 +16,7 @@ class IntegrationTest(PrawTest):
 
     def setup(self):
         """Setup runs before all test cases."""
+        self._overrode_reddit_setup = True
         self.setup_reddit()
         self.setup_betamax()
 
@@ -29,11 +32,32 @@ class IntegrationTest(PrawTest):
         # Require tests to explicitly disable read_only mode.
         self.reddit.read_only = True
 
+        pytest.set_up_record = self.set_up_record  # used in conftest.py
+
     def setup_reddit(self):
+        self._overrode_reddit_setup = False
+
+        self._session = requests.Session()
+
         self.reddit = Reddit(
+            requestor_kwargs={"session": self._session},
             client_id=pytest.placeholders.client_id,
             client_secret=pytest.placeholders.client_secret,
             password=pytest.placeholders.password,
             user_agent=pytest.placeholders.user_agent,
             username=pytest.placeholders.username,
         )
+
+    def set_up_record(self):
+        if not self._overrode_reddit_setup:
+            if (
+                pytest.placeholders.refresh_token
+                != "placeholder_refresh_token"
+            ):
+                self.reddit = Reddit(
+                    requestor_kwargs={"session": self._session},
+                    client_id=pytest.placeholders.client_id,
+                    client_secret=pytest.placeholders.client_secret,
+                    user_agent=pytest.placeholders.user_agent,
+                    refresh_token=pytest.placeholders.refresh_token,
+                )