+++ tests/unit/util/test_cache.py
@@ -22,12 +22,20 @@ class TestCachedProperty(UnitTest):
         assert "nine" not in klass.__dict__
         assert klass.nine == 9
         assert "nine" in klass.__dict__
+        assert "ten" not in klass.__dict__
+        assert klass.ten == 10
+        assert "ten" in klass.__dict__
 
     def test_repr(self):
         klass = self.Klass()
         assert repr(klass.nine) == "9"
 
         property_repr = repr(self.Klass.nine)
+        assert property_repr.startswith("<cachedproperty <function")
+
+        assert repr(klass.ten) == "10"
+
+        property_repr = repr(self.Klass.ten)
         assert property_repr.startswith("<cachedproperty <function")
 
     def test_doc(self):